requirements.txt
sqlalchemy
pandas
fastapi
pydantic
python-dotenv
psycopg2-binary